requirements.txt
sentence-transformers
faiss-cpu
numpy
pdfplumber
python-docx
requests